Checking for user's permissions on back-end (not finished)

+++ final/actions/member/member_delete_user.php
@@ -9,29 +9,22 @@ $response = array();
 $data = json_decode(stripslashes($_POST['data']));
 
 $password = $data[0];
-$usernames = $data[1];
+$username = $data[1];
 $currentUser = getMemberByUsername($_SESSION['username']);
 
-$response['message']="All users were successfully deleted";
-$response['users']=" ";
-
-//if(strcmp(str_replace(' ', '',$currentUser['hashed_pass']), sha1($password)) != 0){ //check for password
-//$response['message'] = "Password does not match";
-
-//} else if($currentUser['privilege_level'] != "Administrator"){ //check for permissions
-//$response['message'] = "User does not have permissions for that";
-
-//} else{
-session_unset();
-session_destroy();
-global $conn;
-$stmt = $conn->prepare("DELETE FROM member WHERE username = ?");
-if(!$stmt->execute(array($usernames))){ //if operation went wrong
-    $response['message'] = "Some users may have not been deleted: ";
-    $response['users'] = $response['users'].$user." ";
+if(strcmp($username,$currentUser['username'])!=0)
+    $response="There seems to be a problem with your username. Contact an Administrator for more information.";
+else if(!password_verify($password, substr($currentUser['hashed_pass'],0,60)))
+    $response="Password was not correct.";
+else {
+    session_unset();
+    session_destroy();
+    global $conn;
+    $stmt = $conn->prepare("DELETE FROM member WHERE username = ?");
+    if(!$stmt->execute(array($username))){ //if operation went wrong
+        $response = "Something went wrong while deleting you. Contact an Administrator for more information.";
+    } else $response = "You were successfully deleted. We're sorry to see you go...";
 }
-//}
-
 
 echo json_encode($response);
